backend/requirements.txt
feedparser
sumy
nltk
lxml
lxml_html_clean
numpy
requests
tinycss2
supabase
python-dotenv
google-generativeai
newspaper3k
beautifulsoup4
